chore: Reorganise and clean code
