--- src/components/miscellaneous/Footer.jsx
@@ -1,19 +1,0 @@
-import React from "react";
-
-const Footer = () => (
-	<footer>
-		<span>Nathanaël Houn © 2021</span>
-		<span>
-			Favicon made by{" "}
-			<a href="https://www.flaticon.com/free-icon/calendar_2693507" title="srip">
-				srip
-			</a>{" "}
-			from{" "}
-			<a href="https://www.flaticon.com/" title="Flaticon">
-				www.flaticon.com
-			</a>
-		</span>
-	</footer>
-);
-
-export default Footer;
