--- src/components/miscellaneous/Loading.scss
@@ -1,3 +1,0 @@
-.loading {
-	text-align: center;
-}
