+++ src/utils/useLocalStorage.js
@@ -16,8 +16,8 @@ export function useLocalStorage(key, initialValue) {
 		}
 	});
 
-	// Return a wrapped version of useState's setter function that ...
-	// ... persists the new value to localStorage.
+	// Return a wrapped version of useState's setter function that
+	// persists the new value to localStorage.
 	const setValue = (value) => {
 		try {
 			// Allow value to be a function so we have same API as useState
